Fix typecast of pointer to u32. Use ptrdiff_t instead, which works on both 32 and 64 bit cpus.
From: Chad Reese <kreese@caviumnetworks.com>
Date: 2010-10-04 10:19:14 -0700

git-svn-id: http://casvn/svn/octeon/bdk/trunk@53591 c83883e8-9f28-0410-baf1-f2025811e16a

+++ eLua/src/common.c
@@ -478,11 +478,11 @@ extern char end[];
 void* platform_get_first_free_ram( unsigned id )
 {
   void* mstart[] = MEM_START_ADDRESS;
-  u32 p;
+  ptrdiff_t p;
 
   if( id >= sizeof( mstart ) / sizeof( void* ) )
     return NULL;
-  p = ( u32 )mstart[ id ];
+  p = ( ptrdiff_t )mstart[ id ];
   if( p & ( MIN_ALIGN - 1 ) )
     p = ( ( p >> MIN_ALIGN_SHIFT ) + 1 ) << MIN_ALIGN_SHIFT;
   return ( void* )p;
@@ -491,11 +491,11 @@ void* platform_get_first_free_ram( unsigned id )
 void* platform_get_last_free_ram( unsigned id )
 {
   void* mend[] = MEM_END_ADDRESS;
-  u32 p;
+  ptrdiff_t p;
 
   if( id >= sizeof( mend ) / sizeof( void* ) )
     return NULL;
-  p = ( u32 )mend[ id ];
+  p = ( ptrdiff_t )mend[ id ];
   if( p & ( MIN_ALIGN - 1 ) )
     p = ( ( p >> MIN_ALIGN_SHIFT ) - 1 ) << MIN_ALIGN_SHIFT;
   return ( void* )p;
